Modified Spigot patch to only ignore Mineshaft structure data.
With the introduction of 1.6.4, Mojang added extra structure info to be
saved in NBT in preparation for 1.7. Mineshafts had a ton of data added to them
which caused mineshaft.dat to grow quickly. If this file was a few
MB+ in size, any save would cause threads to get blocked and lag the
server as it generated the new mineshaft.dat file.
To work around this issue, all structure saving info for mineshafts will be
ignored by default. If for some reason you need this turned back on, reenable
save-mineshaft-structure-info in spigot.yml. This patch will be removed in
1.7 as it will no longer be needed.