Prevent EntityItemFrame from being serialized. Fixes #474, #506, #520
Due to a bug in Gson :
(https://code.google.com/p/google-gson/issues/detail?id=440),
a stackoverflow can occur when gson attempts to resolve a field of
a class that points to itself. As a temporary workaround, we will
prevent serialization for this object until the bug is fixed.
This fixes EE3's serialization of ItemStack.

Thanks to pahimar for helping me track down this issue.