Update to 1.7.2 and migrate to Gradle
Still not ready but the bulk of update is finished. Big thanks to AbrarSyed for helping us migrate to Gradle and rest of Forge team for
allowing us to use their tools.

MCPC dev environment has been completely redone from scratch thanks to
AbrarSyed. The environment is setup identical to forge with Clean
being Forge/FML. The general process goes like this :

1. FML patches Vanilla
2. Forge patches FML
3. MCPC patches Forge
4. Distribution jars are built

We still have some work to do but 1.7 should be ready soon!