Fixed chat issues with mods such as Psychedelicraft.
Fixes #1041, #1245, #1343

When sending messages back to clients, CraftBukkit would use the
ChatComponentText object in order to send less data. While this is good,
it breaks compatibility with mods that expects a ChatComponentTranslation
object. To workaround this compatibility issue, I have modified
CraftChatMessage to send a ChatComponentTranslation object in order to
maintain compatibility with Forge.

* Added support for running plugins in dev environment