Ignore event damage check for EntityItemFrames. Fixes #1488
When a player attempts to remove an item from an EntityItemFrame, CB calls
its handleNonLivingEntityDamageEvent method so plugins have the ability to
prevent the removal. However, this method also checks to see if the event
damage is 0 and if so cancels the removal which in this case causes the bug to
happen.